fix: change notification event handler WISH-458
